backend/pyproject.toml
[project]
name = "jobbr-backend"
version = "0.1.0"
description = "Jobbr backend API"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "langchain-anthropic>=0.3.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
    "resend>=2.0.0",
    "PyJWT>=2.8.0",
    "pgvector>=0.3.0",
    "voyageai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
